Rename and move particle distribution file
